svn:revision:68145
svn:branch:brep-debug
svn:account:n_reed

+++ src/conv/iges/readtree.c
@@ -44,7 +44,7 @@ Readtree(mat_t *matp)
 	    RT_TREE_INIT(ptr);
 	    ptr->tr_l.tl_op = OP_DB_LEAF;
 	    k = ((-op)-1)/2;
-	    if (k < 0 || k >= totentities) {
+	    if (k < 0 || (size_t)k >= totentities) {
 		bu_log("Readtree(): pointer in tree is out of bounds (%d)\n", -op);
 		return (union tree *)NULL;
 	    }
